@@ -12,7 +12,6 @@ import static java.util.Objects.requireNonNull;
  * Represents an immutable json number of type long.
  */
 public final class JsLong extends JsNumber implements Comparable<JsLong> {
-    public static final int TYPE_ID = 7;
 
     /**
      * prism between the sum type JsValue and JsLong
@@ -50,10 +49,6 @@ public final class JsLong extends JsNumber implements Comparable<JsLong> {
         return new JsLong(n);
     }
 
-    @Override
-    public int id() {
-        return TYPE_ID;
-    }
 
     @Override
     public boolean isLong() {
@@ -94,8 +89,7 @@ public final class JsLong extends JsNumber implements Comparable<JsLong> {
     public boolean equals(final Object that) {
         if (this == that) return true;
         if (that == null) return false;
-        if (!(that instanceof JsNumber)) return false;
-        JsNumber number = (JsNumber) that;
+        if (!(that instanceof JsNumber number)) return false;
         if (number.isLong()) return value == number.toJsLong().value;
         if (number.isInt()) return intEquals(number.toJsInt());
         if (number.isBigInt()) return bigIntEquals(number.toJsBigInt());
