@@ -3,7 +3,7 @@ package jsonvalues;
 /**
  * Represents an immutable json number. It's a marker interface for the types {@link JsInt}, {@link JsLong}, {@link JsDouble}, {@link JsBigInt} and {@link JsBigDec}
  */
-public abstract  class JsNumber extends JsPrimitive  {
+public abstract sealed class JsNumber extends JsPrimitive permits JsBigDec, JsBigInt, JsDouble, JsInt, JsLong {
 
     @Override
     public JsPrimitive toJsPrimitive() {
