@@ -19,7 +19,7 @@ final class JsLongReader extends AbstractReader {
                         ) throws IOException {
         long value = NumberConverter.deserializeLong(reader);
         Optional<JsError> result = fn.apply(value);
-        if (!result.isPresent()) return JsLong.of(value);
+        if (result.isEmpty()) return JsLong.of(value);
         throw JsParserException.reasonAt(ParserErrors.JS_ERROR_2_STR.apply(result.get()),
                                          reader.getPositionInStream()
                                         );
