@@ -34,7 +34,7 @@ class JsValueReader extends AbstractReader {
                          ) throws IOException {
         JsValue value = value(reader);
         Optional<JsError> result = fn.apply(value);
-        if (!result.isPresent()) return value;
+        if (result.isEmpty()) return value;
         throw JsParserException.reasonAt(ParserErrors.JS_ERROR_2_STR.apply(result.get()),
                                          reader.getPositionInStream()
                                         );
