@@ -22,70 +22,37 @@ final class JsValueWritter {
                    final JsValue value
                   ) {
 
-        switch (value.id()) {
-            case JsBool.TYPE_ID: {
-                writer.writeAscii(Boolean.toString(value.toJsBool().value));
-                break;
-            }
-            case JsNull.TYPE_ID: {
-                writer.writeNull();
-                break;
-            }
-            case JsStr.TYPE_ID: {
-                writer.writeString(value.toJsStr().value);
-                break;
-            }
-            case JsObj.TYPE_ID: {
-                objectSerializer.write(writer,
-                                       value.toJsObj()
-                                      );
-                break;
-            }
-            case JsArray.TYPE_ID: {
-                arraySerializer.write(writer,
-                                      value.toJsArray()
-                                     );
-                break;
-            }
-            case JsDouble.TYPE_ID:
-            case JsBigDec.TYPE_ID: {
-                NumberConverter.serialize(value.toJsBigDec().value,
-                                          writer
-                                         );
-                break;
-            }
-            case JsBigInt.TYPE_ID: {
-                writer.writeAscii(value.toJsBigInt().value
-                                          .toString());
-
-                break;
-            }
-            case JsLong.TYPE_ID: {
-                NumberConverter.serialize(value.toJsLong().value,
-                                          writer
-                                         );
-                break;
-            }
-            case JsInt.TYPE_ID: {
-                NumberConverter.serialize(value.toJsInt().value,
-                                          writer
-                                         );
-                break;
-            }
-
-            case JsBinary.TYPE_ID: {
-                byte[] xs = value.toJsBinary().value;
+        switch (value) {
+            case JsBool bool -> writer.writeAscii(Boolean.toString(bool.value));
+            case JsNull ignored -> writer.writeNull();
+            case JsStr str -> writer.writeString(str.value);
+            case JsObj obj -> objectSerializer.write(writer,
+                                                     obj
+                                                    );
+            case JsArray arr -> arraySerializer.write(writer,
+                                                      arr
+                                                     );
+            case JsDouble d -> NumberConverter.serialize(d.toJsBigDec().value,
+                                                         writer
+                                                        );
+            case JsBigDec bd -> NumberConverter.serialize(bd.value,
+                                                          writer
+                                                         );
+            case JsBigInt bi -> writer.writeAscii(bi.value.toString());
+            case JsLong l -> NumberConverter.serialize(l.value,
+                                                       writer
+                                                      );
+            case JsInt i -> NumberConverter.serialize(i.value,
+                                                      writer
+                                                     );
+            case JsBinary b -> {
+                byte[] xs = b.value;
                 if (xs.length == 0) writer.writeString("");
                 else writer.writeBinary(xs);
-                break;
             }
-            case JsInstant.TYPE_ID: {
-                writer.writeString(value.toJsInstant().value.toString());
-                break;
-            }
-            default:
-                throw new IllegalStateException("JsValue.id() not considered. Default branch of a switch statement was executed.");
-
+            case JsInstant instant -> writer.writeString(instant.value.toString());
+            case JsNothing ignored -> writer.writeString("");
         }
+
     }
 }
